+++ Package.swift
@@ -91,7 +91,7 @@ let package = Package(
 			targets:["RAW_sha512"])
 	],
 	dependencies: [
-		.package(url:"https://github.com/apple/swift-syntax.git", "600.0.1"..<"620.0.0"),
+		.package(url:"https://github.com/apple/swift-syntax.git", "602.0.0"..<"603.0.0"),
 		.package(url:"https://github.com/apple/swift-log.git", "1.0.0"..<"2.0.0")
 	],
 	targets: [