+++ Sources/RAW_macros/RAW_staticbuff_macro.swift
@@ -1036,7 +1036,7 @@ public struct RAW_staticbuff_concat_macro:MemberMacro, ExtensionMacro {
 		""")]
 	}
 
-	public static func expansion(of node:SwiftSyntax.AttributeSyntax, providingMembersOf declaration: some SwiftSyntax.DeclGroupSyntax, in context: some SwiftSyntaxMacros.MacroExpansionContext) throws -> [SwiftSyntax.DeclSyntax] {
+	public static func expansion(of node:SwiftSyntax.AttributeSyntax, providingMembersOf declaration: some SwiftSyntax.DeclGroupSyntax, conformingTo protocols:[TypeSyntax], in context: some SwiftSyntaxMacros.MacroExpansionContext) throws -> [SwiftSyntax.DeclSyntax] {
 		
 		guard let (asStruct, typeTokens, varNamesAndTypes, isCompareOverridden) = Self.determineIfUsageCompliant(declaration:declaration, node:node, context:context, addDiagnostics:false) else {
 			return []